% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{dimReduc}
\alias{dimReduc}
\title{dimReduc}
\usage{
dimReduc(threshold, count_mat, is_labeled, clusters, ...)
}
\arguments{
\item{threshold}{}

\item{count_mat}{}

\item{is_labeled}{}

\item{clusters}{}

\item{...}{Additional parameters to pass to rpca}
}
\description{
dimReduc
}
