% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{find_hv_genes}
\alias{find_hv_genes}
\title{find_hv_genes}
\usage{
find_hv_genes(count, I)
}
\arguments{
\item{count}{PARAM_DESCRIPTION}

\item{I}{PARAM_DESCRIPTION}

\item{J}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
