% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{find_va_genes}
\alias{find_va_genes}
\title{find_va_genes}
\usage{
find_va_genes(parslist, subcount)
}
\arguments{
\item{parslist}{PARAM_DESCRIPTION}

\item{subcount}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
