% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mix_parameters.R
\name{fn}
\alias{fn}
\title{fn}
\usage{
fn(alpha, target)
}
\arguments{
\item{alpha}{PARAM_DESCRIPTION}

\item{target}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
root-finding equation
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
