% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mix_parameters.R
\name{get_mix}
\alias{get_mix}
\title{get_mix}
\usage{
get_mix(xdata, point)
}
\arguments{
\item{xdata}{PARAM_DESCRIPTION}

\item{point}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
estimate parameters in the mixture distribution
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
