% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mix_parameters.R
\name{get_mix_parameters}
\alias{get_mix_parameters}
\title{get_mix_parameters}
\usage{
get_mix_parameters(count, point = log10(1.01))
}
\arguments{
\item{count}{PARAM_DESCRIPTION}

\item{point}{PARAM_DESCRIPTION, Default: log10(1.01)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
