% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{imputation_model}
\alias{imputation_model}
\title{Title}
\usage{
imputation_model(count, point, drop_threshold = 0.5, Kcluster = 10,
  labeled = FALSE, cell_labels = NULL)
}
\arguments{
\item{count}{}

\item{point}{}

\item{drop_threshold}{}

\item{Kcluster}{}

\item{labeled}{}

\item{cell_labels}{}
}
\description{
Title
}
