% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{imputation_model8}
\alias{imputation_model8}
\title{imputation_model8}
\usage{
imputation_model8(count, labeled, point, drop_thre = 0.5,
  Kcluster = 10)
}
\arguments{
\item{count}{PARAM_DESCRIPTION}

\item{labeled}{PARAM_DESCRIPTION}

\item{point}{PARAM_DESCRIPTION}

\item{drop_thre}{PARAM_DESCRIPTION, Default: 0.5}

\item{Kcluster}{PARAM_DESCRIPTION, Default: 10}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
