% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{imputation_wlabel_model8}
\alias{imputation_wlabel_model8}
\title{imputation_wlabel_model8}
\usage{
imputation_wlabel_model8(count, labeled, cell_labels = NULL, point,
  drop_thre, Kcluster = NULL)
}
\arguments{
\item{count}{PARAM_DESCRIPTION}

\item{labeled}{PARAM_DESCRIPTION}

\item{cell_labels}{PARAM_DESCRIPTION, Default: NULL}

\item{point}{PARAM_DESCRIPTION}

\item{drop_thre}{PARAM_DESCRIPTION}

\item{Kcluster}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
