% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{impute_nnls}
\alias{impute_nnls}
\title{impute_nnls}
\usage{
impute_nnls(Ic, cellid, subcount, droprate, geneid_drop, geneid_obs, nbs,
  distc)
}
\arguments{
\item{Ic}{PARAM_DESCRIPTION}

\item{cellid}{PARAM_DESCRIPTION}

\item{subcount}{PARAM_DESCRIPTION}

\item{droprate}{PARAM_DESCRIPTION}

\item{geneid_drop}{PARAM_DESCRIPTION}

\item{geneid_obs}{PARAM_DESCRIPTION}

\item{nbs}{PARAM_DESCRIPTION}

\item{distc}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
