% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{impute_values}
\alias{impute_values}
\title{impute_values}
\usage{
impute_values(cellid, num_cells, num_genes, subcount, droprate, dropouts,
  dist_list, non_dropouts)
}
\arguments{
\item{cellid}{}

\item{num_cells}{}

\item{num_genes}{}

\item{subcount}{}

\item{droprate}{}

\item{dropouts}{}

\item{dist_list}{}

\item{non_dropouts}{}
}
\description{
impute_values
}
