% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_count.R
\name{read_count}
\alias{read_count}
\title{read_count_data}
\usage{
read_count(countdata, type, genelen)
}
\arguments{
\item{countdata}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{genelen}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
