% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimpute.R
\name{scimpute}
\alias{scimpute}
\title{FUNCTION_TITLE}
\usage{
scimpute(datExpr, type = "count", drop_thre = 0.5, Kcluster = NULL,
  labels = NULL, genelen = NULL, plan = "multiprocess")
}
\arguments{
\item{datExpr}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: 'count'}

\item{drop_thre}{PARAM_DESCRIPTION, Default: 0.5}

\item{Kcluster}{PARAM_DESCRIPTION, Default: NULL}

\item{labels}{PARAM_DESCRIPTION, Default: NULL}

\item{genelen}{PARAM_DESCRIPTION, Default: NULL}

\item{plan}{PARAM_DESCRIPTION, Default: "multiprocess"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
