% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mix_parameters.R
\name{update_gmm_pars}
\alias{update_gmm_pars}
\title{update_gmm_pars}
\usage{
update_gmm_pars(x, wt)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{wt}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
update parameters in gamma distribution
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
